Adventure Game
This is still a work in progress for whenever I find the time. This is
essentially a simple text based game that asks for your name, desired
race, and desired class. You then walk through the dungeon and attack
monsters as they randomly pop up. The overall goal is to get to level
10.